+++ src/op/iter_all_edges.rs
@@ -15,9 +15,15 @@
 extern crate alloc;
 
 use {
-    alloc::collections::BTreeSet,
+    alloc::collections::{
+        BTreeMap,
+        BTreeSet,
+    },
     core::hash::BuildHasher,
-    std::collections::HashSet,
+    std::collections::{
+        HashMap,
+        HashSet,
+    },
 };
 
 /// A trait to iterate over all edges in an unweighted directed graph
@@ -55,6 +61,121 @@ use {
 pub trait IterAllEdges {
     /// Returns an iterator over all edges in a graph.
     fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)>;
+}
+
+impl IterAllEdges for Vec<Vec<usize>> {
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .enumerate()
+            .flat_map(|(s, vec)| vec.iter().map(move |t| (s, *t)))
+    }
+}
+
+impl IterAllEdges for Vec<BTreeSet<usize>> {
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .enumerate()
+            .flat_map(|(s, set)| set.iter().map(move |t| (s, *t)))
+    }
+}
+
+impl<H> IterAllEdges for Vec<HashSet<usize, H>>
+where
+    H: BuildHasher,
+{
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .enumerate()
+            .flat_map(|(s, set)| set.iter().map(move |t| (s, *t)))
+    }
+}
+
+impl IterAllEdges for [Vec<usize>] {
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .enumerate()
+            .flat_map(|(s, vec)| vec.iter().map(move |t| (s, *t)))
+    }
+}
+
+impl IterAllEdges for [BTreeSet<usize>] {
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .enumerate()
+            .flat_map(|(s, set)| set.iter().map(move |t| (s, *t)))
+    }
+}
+
+impl<H> IterAllEdges for [HashSet<usize, H>]
+where
+    H: BuildHasher,
+{
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .enumerate()
+            .flat_map(|(s, set)| set.iter().map(move |t| (s, *t)))
+    }
+}
+
+impl<const V: usize> IterAllEdges for [Vec<usize>; V] {
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .enumerate()
+            .flat_map(|(s, vec)| vec.iter().map(move |t| (s, *t)))
+    }
+}
+
+impl<const V: usize> IterAllEdges for [BTreeSet<usize>; V] {
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .enumerate()
+            .flat_map(|(s, set)| set.iter().map(move |t| (s, *t)))
+    }
+}
+
+impl<const V: usize, H> IterAllEdges for [HashSet<usize, H>; V]
+where
+    H: BuildHasher,
+{
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .enumerate()
+            .flat_map(|(s, set)| set.iter().map(move |t| (s, *t)))
+    }
+}
+
+impl IterAllEdges for BTreeMap<usize, Vec<usize>> {
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .flat_map(|(s, vec)| vec.iter().map(move |t| (*s, *t)))
+    }
+}
+
+impl IterAllEdges for BTreeMap<usize, BTreeSet<usize>> {
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .flat_map(|(s, set)| set.iter().map(move |t| (*s, *t)))
+    }
+}
+
+impl<H> IterAllEdges for HashMap<usize, Vec<usize>, H>
+where
+    H: BuildHasher,
+{
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .flat_map(|(s, vec)| vec.iter().map(move |t| (*s, *t)))
+    }
+}
+
+impl<H> IterAllEdges for HashMap<usize, HashSet<usize, H>, H>
+where
+    H: BuildHasher,
+{
+    fn iter_all_edges(&self) -> impl Iterator<Item = (usize, usize)> {
+        self.iter()
+            .flat_map(|(s, set)| set.iter().map(move |t| (*s, *t)))
+    }
 }
 
 impl IterAllEdges for Vec<(usize, usize)> {
@@ -106,7 +227,128 @@ mod tests {
     }
 
     #[test]
-    fn vec() {
+    fn vec_vec() {
+        let graph = vec![vec![1], vec![2], vec![0]];
+
+        assert!(graph
+            .iter_all_edges()
+            .eq([(0, 1), (1, 2), (2, 0)].into_iter()));
+    }
+
+    #[test]
+    fn vec_btree_set() {
+        let graph = vec![
+            BTreeSet::from([1]),
+            BTreeSet::from([2]),
+            BTreeSet::from([0]),
+        ];
+
+        test_iter_all_edges_unstable!(graph);
+    }
+
+    #[test]
+
+    fn vec_hash_set() {
+        let graph = vec![HashSet::from([1]), HashSet::from([2]), HashSet::from([0])];
+
+        test_iter_all_edges_unstable!(graph);
+    }
+
+    #[test]
+    fn slice_vec() {
+        let graph: &[Vec<usize>] = &[vec![1], vec![2], vec![0]];
+
+        assert!(graph
+            .iter_all_edges()
+            .eq([(0, 1), (1, 2), (2, 0)].into_iter()));
+    }
+
+    #[test]
+    fn slice_btree_set() {
+        let graph: &[BTreeSet<usize>] = &[
+            BTreeSet::from([1]),
+            BTreeSet::from([2]),
+            BTreeSet::from([0]),
+        ];
+
+        test_iter_all_edges_unstable!(graph);
+    }
+
+    #[test]
+    fn slice_hash_set() {
+        let graph: &[HashSet<usize>] =
+            &[HashSet::from([1]), HashSet::from([2]), HashSet::from([0])];
+
+        test_iter_all_edges_unstable!(graph);
+    }
+
+    #[test]
+    fn arr_vec() {
+        let graph = [vec![1], vec![2], vec![0]];
+
+        assert!(graph
+            .iter_all_edges()
+            .eq([(0, 1), (1, 2), (2, 0)].into_iter()));
+    }
+
+    #[test]
+    fn arr_btree_set() {
+        let graph = [
+            BTreeSet::from([1]),
+            BTreeSet::from([2]),
+            BTreeSet::from([0]),
+        ];
+
+        test_iter_all_edges_unstable!(graph);
+    }
+
+    #[test]
+    fn arr_hash_set() {
+        let graph = [HashSet::from([1]), HashSet::from([2]), HashSet::from([0])];
+
+        test_iter_all_edges_unstable!(graph);
+    }
+
+    #[test]
+    fn btree_map_vec() {
+        let graph = BTreeMap::from([(0, vec![1]), (1, vec![2]), (2, vec![0])]);
+
+        assert!(graph
+            .iter_all_edges()
+            .eq([(0, 1), (1, 2), (2, 0)].into_iter()));
+    }
+
+    #[test]
+    fn btree_map_btree_set() {
+        let graph = BTreeMap::from([
+            (0, BTreeSet::from([1])),
+            (1, BTreeSet::from([2])),
+            (2, BTreeSet::from([0])),
+        ]);
+
+        test_iter_all_edges_unstable!(graph);
+    }
+
+    #[test]
+    fn hash_map_vec() {
+        let graph = HashMap::from([(0, vec![1]), (1, vec![2]), (2, vec![0])]);
+
+        test_iter_all_edges_unstable!(graph);
+    }
+
+    #[test]
+    fn hash_map_hash_set() {
+        let graph = HashMap::from([
+            (0, HashSet::from([1])),
+            (1, HashSet::from([2])),
+            (2, HashSet::from([0])),
+        ]);
+
+        test_iter_all_edges_unstable!(graph);
+    }
+
+    #[test]
+    fn vec_tuple() {
         let graph = vec![(0, 1), (1, 2), (2, 0)];
 
         assert!(graph
@@ -115,7 +357,7 @@ mod tests {
     }
 
     #[test]
-    fn slice() {
+    fn slice_tuple() {
         let graph: &[(usize, usize)] = &[(0, 1), (1, 2), (2, 0)];
 
         assert!(graph
@@ -124,7 +366,7 @@ mod tests {
     }
 
     #[test]
-    fn arr() {
+    fn arr_tuple() {
         let graph = [(0, 1), (1, 2), (2, 0)];
 
         assert!(graph