+++ src/op/mod.rs
@@ -115,6 +115,7 @@ pub mod count_all_edges;
 pub mod count_all_vertices;
 pub mod edge_weight;
 pub mod indegree;
+pub mod is_balanced;
 pub mod is_edge;
 pub mod is_simple;
 pub mod iter_all_edges;
@@ -132,6 +133,7 @@ pub use {
     count_all_vertices::CountAllVertices,
     edge_weight::EdgeWeight,
     indegree::Indegree,
+    is_balanced::IsBalanced,
     is_edge::IsEdge,
     is_simple::IsSimple,
     iter_all_edges::IterAllEdges,